+++ FirstCPPApplication/enemies/troll.cpp
@@ -64,7 +64,7 @@ void Troll::championize()
     this->representation->setFGColor(TCODColor::green+TCODColor::green+TCODColor::darkYellow, true, false, true);
     this->attrs->health->current_val+=this->attrs->health->current_val;
     this->attrs->health->max_val+=this->attrs->health->max_val;
-    this->xp_value= floor(this->xp_value*1.5);
+    this->xp_value= (int)floor(this->xp_value*1.5);
 }
 
 void Troll::Die()
