+++ FirstCPPApplication/utils.cpp
@@ -188,6 +188,6 @@ std::string char_to_str(char chr)
 
 int get_euclidean_distance(int x1, int y1, int x2, int y2)
 {
-    return std::sqrt(std::pow(x2-x1, 2.0f)+std::pow(y2-y1, 2.0f));
+    return (int)std::sqrt((int)std::pow(x2-x1, 2.0f)+(double)(int)std::pow(y2-y1, 2.0f));
 };
